Refatoração das classes de usuário e aplicação de herança
